Move stat funcs into a stats.js file
